Refactor NextAuth configuration into separate lib/auth.ts file
