+++ src/app/api/auth/[...nextauth]/route.ts
@@ -1,84 +1,6 @@
 import NextAuth from 'next-auth'
-import CredentialsProvider from 'next-auth/providers/credentials'
-import connectDB from '@/lib/mongodb'
-import User from '@/models/User'
-import type { NextAuthOptions } from 'next-auth'
-
-const authOptions: NextAuthOptions = {
-  providers: [
-    CredentialsProvider({
-      name: 'credentials',
-      credentials: {
-        email: { label: 'Email', type: 'email' },
-        password: { label: 'Password', type: 'password' }
-      },
-      async authorize(credentials) {
-        if (!credentials?.email || !credentials?.password) {
-          throw new Error('Email und Passwort sind erforderlich')
-        }
-
-        try {
-          await connectDB()
-          
-          const user = await User.findOne({ 
-            email: credentials.email,
-            isActive: true 
-          }).select('+password')
-
-          if (!user) {
-            throw new Error('Ungültige Anmeldedaten')
-          }
-
-          const isPasswordValid = await user.comparePassword(credentials.password)
-          
-          if (!isPasswordValid) {
-            throw new Error('Ungültige Anmeldedaten')
-          }
-
-          return {
-            id: user._id.toString(),
-            email: user.email,
-            firstName: user.firstName,
-            lastName: user.lastName,
-            role: user.role,
-          }
-        } catch (error) {
-          console.error('Auth error:', error)
-          throw new Error('Anmeldung fehlgeschlagen')
-        }
-      }
-    })
-  ],
-  callbacks: {
-    async jwt({ token, user }) {
-      if (user) {
-        token.role = user.role
-        token.firstName = user.firstName
-        token.lastName = user.lastName
-      }
-      return token
-    },
-    async session({ session, token }) {
-      if (token) {
-        session.user.id = token.sub!
-        session.user.role = token.role as string
-        session.user.firstName = token.firstName as string
-        session.user.lastName = token.lastName as string
-      }
-      return session
-    }
-  },
-  pages: {
-    signIn: '/login',
-    error: '/login',
-  },
-  session: {
-    strategy: 'jwt',
-    maxAge: 30 * 24 * 60 * 60, // 30 days
-  },
-  secret: process.env.NEXTAUTH_SECRET,
-}
+import { authOptions } from '@/lib/auth'
 
 const handler = NextAuth(authOptions)
 
-export { handler as GET, handler as POST, authOptions }
+export { handler as GET, handler as POST }
